refactored filtering, get_model_properties, index creation tests
