+++ src/marqo/tensor_search/models/index_info.py
@@ -1,14 +1,36 @@
 import pprint
 from typing import NamedTuple, Any, Dict
 from marqo.tensor_search import enums
-from marqo.tensor_search.enums import IndexSettingsField as NsFields
+from marqo.tensor_search.enums import IndexSettingsField as NsField
 from marqo.tensor_search import configs
+from marqo.s2_inference import s2_inference
+from marqo import errors
+from marqo.s2_inference import errors as s2_inference_errors
+
+
+# Can be used outside of this module
+def get_model_properties_from_index_defaults(index_defaults: Dict, model_name: str):
+    """ Gets model_properties from index defaults if available
+    """
+    try:
+        model_properties = index_defaults[NsField.model_properties]
+    except KeyError:
+        try:
+            model_properties = s2_inference.get_model_properties_from_registry(model_name)
+        except s2_inference_errors.UnknownModelError:
+            raise errors.InvalidArgError(
+                f"Could not find model properties for model={model_name}. "
+                f"Please check that the model name is correct. "
+                f"Please provide model_properties if the model is a custom model and is not supported by default")
+    return model_properties
+
 
 class IndexInfo(NamedTuple):
     """
     model_name: name of the ML model used to encode the data
     properties: keys are different index field names, values
         provide info about the properties
+    index_settings: settings for the index
     """
     model_name: str
     properties: dict
@@ -54,6 +76,12 @@ class IndexInfo(NamedTuple):
                     for sub_field, sub_field_props in text_props["properties"].items():
                             text_props_dict[f"{text_field}.{sub_field}"] = sub_field_props
         return text_props_dict
+
+    def get_model_properties(self) -> dict:
+        index_defaults = self.index_settings["index_defaults"]
+        return get_model_properties_from_index_defaults(
+            index_defaults=index_defaults, model_name=self.model_name
+        )
 
     def get_possible_tensor_fields(self) -> set:
         """returns all fields that have are, or have the potential to be, kNN fields
@@ -102,16 +130,16 @@ class IndexInfo(NamedTuple):
         
         """
         ann_default = configs.get_default_ann_parameters()
-        index_ann_defaults = self.index_settings[NsFields.index_defaults].get(NsFields.ann_parameters, {})
+        index_ann_defaults = self.index_settings[NsField.index_defaults].get(NsField.ann_parameters, {})
 
         # index defaults override generic defaults
         ann_params = {
             **ann_default,
             **index_ann_defaults
         }
-        ann_params[NsFields.ann_method_parameters] = {
-            **ann_default[NsFields.ann_method_parameters],
-            **index_ann_defaults.get(NsFields.ann_method_parameters, {})
+        ann_params[NsField.ann_method_parameters] = {
+            **ann_default[NsField.ann_method_parameters],
+            **index_ann_defaults.get(NsField.ann_method_parameters, {})
         }
 
         return ann_params
